requirements.txt
streamlit>=1.20.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.5
rembg>=2.0.30
matplotlib>=3.5.0
onnxruntime>=1.10.0
onnxruntime
os
